+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added support for XFS disk isolator project ID metrics.
+
 ## [1.1.1] - 2018-10-25
 ### Fixed
 - Fixed a regression in v1.1.0 that caused the exporter to crash